[surpass] The first energy term implemented and tested: Excluded volume
 - SurpassEnergy trait defined
 - MoveProposal::moved_cax renamed to MoveProposal::cax (also cay, caz) to simplify excluded_volume_kernel!() macro: since both SurpassAlphaSystem and  MoveProposal have cax, the macro can be used in all energy evaluations
 - test_coords_operations() and test_diatance_evaluation() added to test_surpass_system.rs to ensure the PBC are applied correctly

+++ surpass/src/system.rs
@@ -67,7 +67,7 @@ impl SurpassAlphaSystem {
 
     #[inline(always)]
     pub fn real_to_int(&self, v: f64) -> i32 {
-        (v / self.int_to_real).rem_euclid((i32::MAX as f64) + 1.0) as i32
+        (v / self.int_to_real).rem_euclid((u32::MAX as f64) + 1.0)  as u32 as i32
     }
 
     /// Returns the number of atoms in this system (of all its chains)
